Mostly functional room booking
Emails are coming in correctly with the room information but when looking at the website the availability is not updated so I have to investigate if its actually working by looking at the supabase. Also you can still select dates more than 4 days away, not all the time slots that should be there are there, and availability of rooms is not accounted for.